Rework proof construction and submission
Auto-submit now works correctly. submitFundingProof becomes
constructFundingProof, which returns an array of parameters to submit to
an on-chain contract for a funding proof. This works well, as the intent
is to also support calling the vending machine's
unqualifiedDepositToTbtc method, which uses the same parameters but has
some other data sent in as well. This way we can reuse the same function
for that.