Refactor BitcoinHelpers confirmation handling
There is now a checkForConfirmations function which checks a given
transaction for a certain number of required confirmations. If that
transaction has enough confirmations, it returns the current number of
confirmations. It returns a promise to that final result.

This function passes through to checkForConfirmationsWithClient, which
is now used in waitForConfirmations as well and does the heavy lifting.