Implement Deposit.getRedemptionCost()
This function returns the redemption cost for this deposit.

If the deposit belongs to the vending machine, it adds 1 TBTC to the
deposit's own owner redemption TBTC requirement (since redeeming a
vending machine-owned deposit requires first retrieving the deposit in
exchange for 1 TBTC, then paying any remaining redemption amount as the
owner of the deposit to redeem it).

If the deposit does not belong to the vending machine, this function
directly calls the contract to get its redemption TBTC requirement; note
that the contract will return an error message if the caller does not
own the deposit, provided that the deposit is not at term.