Separate keep public key handling and Bitcoin address
The handling for the public key is now its own promise, and that promise
is used to feed the Bitcoin address promise for the deposit. This gives
us the ability to access the public key independently of the Bitcoin
address from elsewhere; in particular, the redemption flow will need
this for constructing proper signed transactions.