Fill out the Depoist.requestRedemption method
This method does a lot of the early async heavy lifting to actually call
the redemption request functions, and forks based on whether the deposit
is currently owned by the vending machine (`vendingMachine.tbtcToBtc`)
or by the current account (`deposit.requestRedemption`). It continues to
return a Redemption object, which now picks up once the redemption
request has already been registered.