Guard requestRedemption on TBTC balance
If the current account does not have sufficient TBTC balance to redeem,
throw rather than continue down the path of madness. Though the contract
itself would revert the transaction in this case that both (a) costs gas
and (b) leads to a slightly less informative error (though not hugely
so).