getLatestRedemptionDetails moves to Deposit
Deposit now has a way to retrieve an already-in-progress redemption, via
getCurrentRedemption(). Redemption now takes the redemption details in
its constructor, where evailable. Redemption details are retrieved
directly from the transaction when initially requesting a redemption,
and from the chain when hitting getCurrentRedemption(). If they aren't
provided directly to the Redemption constructor, they are looked up from
the chain.