Bump bcoin and drop TypeScript dependencies
TypeScript was a pain, so we're not using it for now. bcoin's tagged
version pulls in leveldb bindings that don't compile on latest Node, so
that's out the window. Anchoring to a bcoin 2.0-pre commit hash for now
since we actually don't use any of the complicated stuff anyway, just
the easy JS paths.