Replace Deposit cancelAutoMonitor with autoMonitor opt-in
The flow of explicitly opting out of auto-monitoring makes less sense in
a world where a deposit can be instantiated after it's already active.
In light of that, switch away from that model altogether and require an
explicit call to an `autoSubmit()` function on the deposit instance to
enable auto-monitoring and funding proof submission.