Construct proofs based on required confirmations
The confirmations count being passed to the SPV proof construction was
the number of confirmations the transaction already had. Instead, this
can be the number of confirmations that are required, which will be less
than or equal to the number of confirmations the transaction has, since
we always wait for the confirmations to be sufficient before
constructing and submitting the proofs.

The biggest upside here is that the constructed proof is the minimum
size it needs to be to establish the necessary facts.