Transpile ES6 with Babel.
We're using class members, which aren't out-of-the-box for many ES6 bundle setups (such as react-scripts). Grrrrrrrrr. TBTC.config is a culprit, eg.

| class TBTC {
>     config/*: TBTCConfig*/;
|
|     constructor(config/*: TBTCConfig*/ = defaultConfig, networkMatchCheck = true) {

There may be others, for now here's a branch that transpiles it. Unfortunately I'm at my wits end, and there's a remaining issue this Babel config and CommonJS modules, which makes importing this package quite ugly.