Move Electrum config to TBTC config object
This instead of reading it from the filesystem, which gets hairy to
manage across browsers and Node. Plus this is part of the configuration
anyway, little reason to shove it in a JSON file in the package.

example.js and resume.js both now inline the config from the JSON file.