Expose auto-submission progress to caller
The deposit auto-submission process involves several distinct phases:

- Waiting for a funding Bitcoin transaction.
- Waiting for that transaction to be confirmed.
- Submitting proof of that transaction to tBTC, and waiting for that
  submission to be mined.

Before, auto-submission was fired off, and the only hook available was
to register an onActive handler. Now, calling Deposit's `autoSubmit`
method returns an `AutoSubmitState` object, with three promises as
properties. These promises, `fundingTransaction`,
`resolvedConfirmations`, and `proofTransaction`, are resolved with each
of the bullets above as they complete. This gives the caller the ability
to react to each stage of the auto-submission process.