Switch JSON to the IBondedECDSAKeep interface
This interface is bundled with the tbtc artifacts, while BondedECDSAKeep
comes from keep-ecdsa. Rather than depend on keep-ecdsa directly, for
now the reference is updated to the interface. In fact, the interface is
all that's necessary for tbtc.js to work correctly, since the ABI is
bound to a per-deposit instance of the contract.