Introduce EthereumHelpers.sendSafely
sendSafely estimates gas and uses that to fill in the gas for the
transaction. If gas estimation fails, the method is attempted as a
`call` in an attempt to discover a better error message. At that point,
two possibilities exist: either the caller can specify the transaction
should still be pushed through with no gas specified, or, by default,
the transaction is not sent and the call or gas estimation error is
propagated up.

All transaction sends in Deposit and Redemption now use
EthereumHelpers.sendSafely.