sendSafely handles unexpected exceptions
sendSafely was assuming the only exception it would get would be an
out-of-gas-or-always-failing-transaction one, but in fact it could see
another, unplanned-for exception thrown. In these cases, the exception
is now rethrown.

Additionally, such an unexpected exception may not necessarily have a
`.message` property, and that is also properly guarded for now.