Work against a numerically sorted version list
Using sort sorts lexicographically which runs us into issues when we
increment pre.x to 10.  Rather than lean into sort -n we had a look at
whether or not the result of jq flatten would return a numerically
ordered list.  It looks like it does, so we stripped the sort.