Stop setting the network
The tbtc dapp is falling over when trying to set this value,
specifically on trying to hit the ./protocol/network path in the bcoin
module.  Current working theory is that this path doesn't get bundled
into the client JS.  Killing it for now to see  if we can get unblocked.