Retry in case of retrieveSignerPubkey error
The `retrieveSignerPubkey` transaction can fail if they occur immediately
after receiving the `PublicKeyPublished` event and a load-balanced
Ethereum client is used. This is because the contract state is
propagated with a delay across all Ethereum nodes. There can be a
situation when the `PublicKeyPublished` is received from one node
but the `retrieveSignerPubkey` is executed against an other one which
isn't aware of the public key yet and will revert the transaction.
To prevent breaking the whole deposit flow, we introduce a retry
mechanism.