Rework TBTC initialization a little
Specifically, move to using EthereumHelpers.isMainnet. Notably, this is
an async function rather than a regular one, so the network match check
has to be performed in the static factory function rather than in the
constructor (which cannot be async).

This moves us overall to using the blessed `web3.eth.getChainId` path to
get the current chain id, rather than the previous
`.currentProvider.chainId` solution which was not part of the public
Web3 API.