Improve Electrum config handling
TBTCConfig now declares that it has an `electrum` property that is an
ElectrumConfig. Until now, that property was in fact a map from various
names (testnet, testnetWS, testnetPublic) to ElectrumConfigs.

BitcoinHelpers now similarly declares that the electrumConfig property
is in fact an ElectrumConfig object. This means the value passed to the
ElectrumClient constructor is now already an ElectrumConfig, rather than
hardcoding BitcoinHelpers to always look up the testnetWS property. This
in turn allows TBTC.withConfig callers to specify Electrum information
even if it's not testnet information, without having to put it under a
testnet key.

Finally, bin/tbtc.js is updated to pass the right object along. Mainnet
configurations are added to bin/tbtc.js's list of available Electrum
configs for future use.