Rework autoSubmit and autoMint for typing purposes
Instead of having an intermediately built-up container for their state,
the containing object is constructed all at once. Additionally, the
AutoSubmitState type now properly reflects the fact that the type is
used for both auto-submission and auto-minting by marking the two fields
that are set in one vs the other as optional.