Add tsc type checking to JS linting
The jsconfig file requires a little bit of adjustment to explicitly
declare some things that VS Code infers implicitly. The output from the
lint is also not beautiful, but it's enough to flag issues for the
committer, which can then be resolved with manual invocation or, better
yet, editor integration with the JS language server.