Adjust the TBTC contracts are defined and accessed
Since contracts are initialized as async promises, their type is
possibly-undefined. Because using the API in an expected way should
never yield undefined contracts, there needs to be a way to access them
that throws if they are undefined, but generally provides access to the
non-optional type. The new accessors on DepositFacotry allow this, with
a helper function, `contractOrBust`, that handles going from a
Contract | undefined type to a Contract type.