EthereumHelpers.sendSafely surfaces revert errors without forceSend
Previously, if forceSend was not specified, the only exception surfaced
by a failing call was the one from `estimateGas`. `estimateGas`,
however, does not report any revert errors, it simply reports that the
submission would run out of gas.

Now, when forceSend is off, a `call` is attempted. If the `call` throws,
that is the exception that bubbles out of `sendSafely`, and it carries
the proper revert reason. If it does not throw, the submission is still
assumed to be problematic since gas estimation failed, and the original
`estimateGas` exception is bubbled out of `sendSafely` instead.