Add outpoint filtering to findScript
findScript, and its helper findWithClient, can now take an outpoint that
will be used as an additional filter to any unspent UTXOs found by
getUnspentToScript. This will allow the redemption flow to ensure that
it is trying to prove a redemption transaction from the correct deposit
in cases where multiple deposits of the same size are being redeemed to
the same address.

Add typing info on Electrum vin output to achieve this.