Bump bcoin to a Keep-forked version
This version anchors bcrypto to version 5.3.0, which appears to be
required in order to successfully npm install on latest macOS.

Several changes are needed to deal with breaking API issues and problems
with babel transcompilation for some of the native bits in bcrypto
5.3.0. These are resolved here, largely through the introduction or
rewriting of BitcoinHelpers functions.

Once upstream is fixed, this can revert to a proper bcoin tagged
version, and the other adjustments will remain relevant.