Add bin/tbtc.js deposit command manager
This is designed to handle a `deposit` command (not yet exposed by
`bin/tbtc.js`). This takes a set of arguments that do not include the
command, and parses them into `new`, `list`, or deposit-specific
subcommands that come after a deposit address. The code is structured as
a set of command parsers which manage argument analysis, and existing
functions mostly taken from the original bin/tbtc.js code.

Notably, output is now standardized via `standardDepositOutput`, which
takes care of tab-delimiting the output.

For clarity, the deposit command help is at the top of the file.