Restructure bin/tbtc.js to use command helpers
This includes both the command helpers and the deposit subcommand
helpers. Much of the deposit handling is removed from bin/tbtc.js, while
a few utilities are included for lot size and supply information lookup.

Additionally, the private key mnemonic and RPC URL can be specified on
the command-line, and debug logging is disabled by default, with a
--debug flag allowing it to be turned back on.