Fix EthereumHelpers.getExistingEvent to return latest
This function was accidentally returning the earliest matching event
instead of the latest matching event, which led to issues when dealing
with a fee-bumped redemption. The latest redemption details would
resolve as the earliest ones, and that would cause a cascading failure
in follow-up functionality since the output and fee data was wrong.