Fallaback to interval past events querying on subscription error
If there is a problem with subscription to events we no longer reject
promise and clear the interval. In case of subscription problem we want
to fallback to the solution where we query past events in an interval.

The change was made after testing tbtc.js `getEvent` function locally.
When subscribing to an event in `once` function a following error was returned:
`The method eth_subscribe does not exist/is not available`. As a result
`clearInterval` function was executed and the fallback defined in
`setInterval` was never executed.