Use pWaitFor instead setInterval
`setInterval` doesn't fully respect async nature of the action that is
called and doesn't block until the action is fulfilled. When a call to
get events takes long time (longer than 3 seconds) with setInterval we're
starting subsequent calls without waiting for the function to resolve on
the first call. `pWaitFor` will wait for each call to resolve
before starting next action.