Updated way to look for past transactions in electrum
Previously we were looking in electrum only for unspent transactions,
which seemed fine for funding but caused problems in redemption. There
are situations on mainnet where a transaction proof has not been
submitted to the deposit yet, but the output has been spent already.
With this change we're looking for the whole history of the script
(including mempool) and filtering the results of our interest.