Filter out failed getTransaction results
In `getTransactionsForScript` we try to get details of each found
transaction. It may happen that `getTransaction` call fails (e.g. when
transaction is not found due to sitting in mempool). In such case we
don't want to abort the whole process and just log an error. Doing this
caused the list to be filled with `undefined` entries. We added
filtering out these `undefined` results so we have a clean list.