Define deployemnt block for contracts
We added a deployedAtBlock property to a contract object. The `Contract`
type now extends web3's Contract Class with a field `deployedAtBlock`.
The field is populated with a value reflecting the block number when
contract was deployed.

For the contracts artifacts containing hash of the transaction when
contract was deployed we use this hash to obtain block number from
chain.
For contracts that are deployed like `Deposit` contract we read an event
of contract creation and store block number as deployment block number
for Deposit and Keep.