Add gasMultiplier to EthereumHelpers.sendSafely
This parameter allows the estimated gas (or the explicitly-specified
gas, if it's higher) to be multiplied by a multiplier that can give some
overhead. It's useful in cases where the transaction in question can
experience some variability in its gas usage based on changing on-chain
conditions that may differ between when gas estimation runs and when the
transaction is mined.