Apply a gas multiplier of 1.4 to deposit creation transaction
This will take the estimate for deposit creation and give it a 20% boost
to account for changes in the sortition pool between gas estimation and
transaction mining.

Note that this is not a guarantee that the 20% boost will be enough to
cover sortition pool changes, but it should absorb most such changes.