Bump web3 to 1.3.0
The core motivator here is to pull in fixes for revert reason handling
in geth >=v1.9.15. Most providers have bumped their geth versions to
v1.9.24 as a dual result of the geth hard fork that resulted from a
dormant bug between 1.9.8 and 1.9.18, and a combined Go/geth panic bug
in consensus code fixed in 1.9.24.