Check for beneficiary *pubs before addresses
The inner content of *pubs can match the address regexp, so switch the
check around. Also, *pubs are accepted even if more than one is
detected, provided that all matched *pubs are the same. This can happen
in cases where the message is a JSON payload of a type that can be
outputted by some tools to include more info than just the *pub.