Refunds handle close/exit event ordering better
Sometimes the `exit` event on the process will fire before the stdout
stream has finished reporting all of its output. In these cases, if the
process completed successfully, the output needs to be allowed to finish
before the public key and signature can be extracted.

The code now waits for the `close` event on the stdout stream, and
throws in cases where the final output cannot be split into public key
and signature correctly.