Fix race condition in keep-ecdsa output handling for refunds
When keep-ecdsa was invoked and exited, sometimes the exit would happen
before the process would finish notifying the JS process of output. The
result would be partial output that could not be parsed correctly. To
fix this, a 'close' event handler was added on the stdout stream when
the process exited. Unfortunately, this broke behavior when stdout had
already been closed, and made the whole process hang.

Whether or not the output has ended is now tracked, and if it has it is
immediately processed on process exit. If not, an event handler is
attached to process the output at that point.