Error handler for Deposit autoSubmit errors
Previously when an error was returned from `provideBTCFundingProof`
transaction in the `autoSubmit` flow an UnhandledPromiseRejectionWarning
was logged, user of the library had no chance to handle this error.

As autoSubmit happens asynchronously we need a way to bubble up the
error. In this proposed approach we catch the error an emit it as an
event. User is required to define error callback with `onError` function.