Allow owner and grant lookup for old staking contract
By default, bin/owner-lookup.js continues to check the new staking
contract for ownership info. However, if used as a library, it now
allows the return of the grant type (as "no grant", "direct grant", and
"managed grant") and allows the caller to substitute the old staking
contract for the `TokenStaking` key of the passed contracts.

To do this, the process for resolving TokenGrantStakes changes
completely. The old staking contract provides no strategy for looking
up an event for the stake delegation using just the operator address,
meaning the owner is needed in order to properly resolve information
about the grant---but this script uses the grant info to resolve
the owner!

Instead, the new approach directly reads the on-chain storage slots and
heuristically determines if a contract is a TokenGrantStake. If it does,
it assumes the fourth slot represents the grant id, and resolves the
grant id in this way.