Notify CI about completion of Node.js workflow
As part of work on RFC-18 a mechanism for managing triggering of
workflows has been created. A part of this mechanism is
the `notify-workflow-action` which will be executed at the end of the
workflow and will notify the release manager repository about build
completion.
This commit introduces the notify-workflow-completed action to
`node.yml` workflow.