Temporarily add `continue-on-error` for step executing tests
The tests run in the `npm test` step has been commented out for some
time, as they need to be updated to pass. After upgrading Node.js to v14
we are getting an error informing about 0 tests executed, which makes
the job fail. We are temporarily allowing for failing of the step, until
the tests get fixed.