Add NPM workflow
This commit adds a NPM workflow similar to what we have in some other
repositories. Workflow will be executed whenever a change in npm
dependencies or in `bin/`, `src/` or `index.js` occurs on `main`. It will
be also possible to execute workflow on demand.
Workflow will update the dependencies to `keep-network` packages and
publish the npm package to the npm registry under bumped-up version
(using `-dev.X` suffix).
Published package will also be tagged with the `development` tag.